+++ src/reuse/_util.py
@@ -17,7 +17,7 @@ from gettext import gettext as _
 from hashlib import sha1
 from os import PathLike
 from pathlib import Path
-from typing import BinaryIO, List, Optional, Set
+from typing import BinaryIO, List, Optional
 
 from boolean.boolean import Expression, ParseError
 from debian.copyright import Copyright
@@ -27,6 +27,7 @@ from . import SpdxInfo
 from ._comment import _all_style_classes
 
 GIT_EXE = shutil.which("git")
+HG_EXE = shutil.which("hg")
 
 _LOGGER = logging.getLogger(__name__)
 _LICENSING = Licensing()
@@ -93,22 +94,7 @@ def execute_command(
     )
 
 
-def find_root() -> Optional[Path]:
-    """Try to find the root of the project from CWD. If none is found, return
-    None.
-    """
-    cwd = Path.cwd()
-    if in_git_repo(cwd):
-        command = [GIT_EXE, "rev-parse", "--show-toplevel"]
-        result = execute_command(command, _LOGGER, cwd=cwd)
-
-        if not result.returncode:
-            path = result.stdout.decode("utf-8")[:-1]
-            return Path(os.path.relpath(path, cwd))
-    return None
-
-
-def find_licenses_directory(root: PathLike = None) -> Optional[Path]:
+def find_licenses_directory(root: PathLike) -> Optional[Path]:
     """Find the licenses directory from CWD or *root*. In the following order:
 
     - LICENSES/ in *root*.
@@ -120,8 +106,6 @@ def find_licenses_directory(root: PathLike = None) -> Optional[Path]:
     The returned LICENSES/ directory NEED NOT EXIST. It may still need to be
     created.
     """
-    if root is None:
-        root = find_root()
     cwd = Path.cwd()
     licenses_path = cwd / "LICENSES"
 
@@ -131,52 +115,6 @@ def find_licenses_directory(root: PathLike = None) -> Optional[Path]:
         licenses_path = cwd
 
     return licenses_path
-
-
-def in_git_repo(cwd: PathLike = None) -> bool:
-    """Is *cwd* inside of a git repository?
-
-    Always return False if git is not installed.
-    """
-    if cwd is None:
-        cwd = Path.cwd()
-
-    if GIT_EXE:
-        command = [GIT_EXE, "status"]
-        result = execute_command(command, _LOGGER, cwd=cwd)
-
-        return not result.returncode
-
-    return False
-
-
-def _all_files_ignored_by_git(root: PathLike) -> Set[Path]:
-    """Return a set of all files ignored by git. If a whole directory is
-    ignored, don't return all files inside of it.
-
-    Return an empty list if git is not installed.
-    """
-    root = Path(root)
-
-    if GIT_EXE:
-        command = [
-            GIT_EXE,
-            "ls-files",
-            "--exclude-standard",
-            "--ignored",
-            "--others",
-            "--directory",
-            # TODO: This flag is unexpected.  I reported it as a bug in Git.
-            # This flag---counter-intuitively---lists untracked directories
-            # that contain ignored files.
-            "--no-empty-directory",
-            # Separate output with \0 instead of \n.
-            "-z",
-        ]
-        result = execute_command(command, _LOGGER, cwd=root)
-        all_files = result.stdout.decode("utf-8").split("\0")
-        return {Path(file_) for file_ in all_files}
-    return set()
 
 
 def decoded_text_from_binary(binary_file: BinaryIO, size: int = None) -> str: